+++ examples/example_pygame.py
@@ -22,9 +22,9 @@ def main():
 
         screen.fill(BLACK)
 
-        with cui.CoshUIRenderer(cui.PygameBackend(screen), cui.DEBUG):
-            with cui.Container(id="main_root", width=cui.FILL, height=cui.FILL, padding=10):
-                cui.Container(id="test", width=100, height=100, style=cui.CoshStyling(background_color=(255, 255, 0)))
+        with cui.CoshUIRenderer(cui.PygameBackend(screen)):
+            with cui.Container(id="main_root", width=cui.FILL, height=cui.FILL, padding=10, align=cui.ALIGN_CENTER, justify=cui.JUSTIFY_CENTER):
+                cui.Container(id="test", width=100, height=100, style=cui.CoshStyling(background_color=(255, 255, 0), border=((255, 100, 100), 10), border_radius=(20, 20, 0, 0)))
 
         if cui.get_signal("test", cui.CLICKED):
             cui.animate("transform_rotation", "test", 45.0, 1.0, "ease_out_bounce")